+++ Classes/Model/Place.php
@@ -4,11 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Model;
 
-use OliverKlee\Oelib\Exception\NotFoundException;
-use OliverKlee\Oelib\Mapper\CountryMapper;
-use OliverKlee\Oelib\Mapper\MapperRegistry;
 use OliverKlee\Oelib\Model\AbstractModel;
-use OliverKlee\Oelib\Model\Country;
 
 /**
  * This class represents a place.
@@ -23,17 +19,12 @@ class Place extends AbstractModel
         return $this->getAsString('title');
     }
 
+    /**
+     * @return string our address, might be empty
+     */
     public function getFullAddress(): string
     {
         return $this->getAsString('address');
-    }
-
-    /**
-     * @return string the ZIP code, might be empty
-     */
-    public function getZip(): string
-    {
-        return $this->getAsString('zip');
     }
 
     /**
@@ -42,26 +33,5 @@ class Place extends AbstractModel
     public function getCity(): string
     {
         return $this->getAsString('city');
-    }
-
-    public function getCountry(): ?Country
-    {
-        $countryCode = $this->getAsString('country');
-        if ($countryCode === '') {
-            return null;
-        }
-
-        try {
-            $country = MapperRegistry::get(CountryMapper::class)->findByIsoAlpha2Code($countryCode);
-        } catch (NotFoundException $exception) {
-            $country = null;
-        }
-
-        return $country;
-    }
-
-    public function hasCountry(): bool
-    {
-        return $this->getCountry() !== null;
     }
 }